Refactor codebase into modular components
This commit restructures the AQIVisualizer project into a more maintainable and comprehensible format by dividing the code into distinct modules. Each module encapsulates specific functionalities, aligning with the Single Responsibility Principle and enhancing the clarity of the codebase.

Changes include:
- `aqi_api.py`: Dedicated to fetching AQI data from the API.
- `image_generator.py`: Handles image generation and text overlay logic.
- `cli.py`: Manages command-line argument parsing.
- `utils.py`: Configures the logging across the application.
- `main.py`: Acts as the entry point, orchestrating the overall program flow.

Additionally, the README.md has been updated to reflect the new project structure, providing clear instructions on setup and execution. This refactor not only improves the current state of the code but also lays a robust foundation for future features and potential contributions.
