+++ utils.py
@@ -1,0 +1,5 @@
+import logging
+
+# Setup logging
+def setup_logging():
+    logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
